Move services to shared folder
