+++ src/app/app.component.html
@@ -1,5 +1,2 @@
 <de-navbar></de-navbar>
-<div>
-</div>
-
 <router-outlet></router-outlet>
